Update to handle newer pytest-cov.
Pytest-cov 2.6 changed how multiple source paths were processed.
Before, all paths were relative to the base path.  Now, the first path
is relative to the base and subsequent paths are relative to the first
path.

The coverage files are all collapsed to a single-named file, so rename
it before combining workspaces in circle-ci so as not to discard data.